Continue chronic warning cleanup across index observability and temporal modules

+++ src/index/product_quantizer.cpp
@@ -130,7 +130,7 @@ ProductQuantizer::Status ProductQuantizer::train(
 
     const size_t expected_dimension = static_cast<size_t>(dimension_);
     for (const auto& vec : training_vectors) {
-        if (static_cast<int>(vec.size()) != expected_dimension) {
+        if (vec.size() != expected_dimension) {
             return Status::Error("Training vector dimension mismatch");
         }
     }
@@ -153,7 +153,7 @@ ProductQuantizer::Status ProductQuantizer::train(
         }
         
         // Validate data size (development safety check)
-        if (static_cast<int>(training_data.size()) != expected_values) {
+        if (training_data.size() != expected_values) {
             return Status::Error("Training data size mismatch during conversion");
         }
         
@@ -177,7 +177,7 @@ ProductQuantizer::Status ProductQuantizer::train(
     // Train each subquantizer independently
     for (int sq = 0; sq < config_.num_subquantizers; ++sq) {
         const size_t sq_index = static_cast<size_t>(sq);
-        if (sq_index >= static_cast<int>(codebooks_.size())) {
+        if (sq_index >= codebooks_.size()) {
             return Status::Error("Subquantizer index out of range");
         }
 
@@ -221,7 +221,7 @@ std::vector<uint8_t> ProductQuantizer::encode(const std::vector<float>& vector) 
         return {};
     }
     
-    if (static_cast<int>(vector.size()) != static_cast<size_t>(dimension_)) {
+    if (vector.size() != static_cast<size_t>(dimension_)) {
         THEMIS_ERROR("ProductQuantizer::encode - Dimension mismatch: {} vs {}",
                      vector.size(), dimension_);
         return {};
@@ -252,7 +252,7 @@ std::vector<uint8_t> ProductQuantizer::encode(const std::vector<float>& vector) 
     // Encode each subvector independently
     for (int sq = 0; sq < config_.num_subquantizers; ++sq) {
         const size_t sq_index = static_cast<size_t>(sq);
-        if (sq_index >= static_cast<int>(codebooks_.size())) {
+        if (sq_index >= codebooks_.size()) {
             THEMIS_ERROR("ProductQuantizer::encode - Subquantizer index out of range");
             return {};
         }
@@ -282,7 +282,7 @@ std::vector<float> ProductQuantizer::decode(const std::vector<uint8_t>& codes) c
         return {};
     }
     
-    if (static_cast<int>(codes.size()) != static_cast<size_t>(config_.num_subquantizers)) {
+    if (codes.size() != static_cast<size_t>(config_.num_subquantizers)) {
         THEMIS_ERROR("ProductQuantizer::decode - Code size mismatch");
         return {};
     }
@@ -312,7 +312,7 @@ std::vector<float> ProductQuantizer::decode(const std::vector<uint8_t>& codes) c
     // Concatenate centroid vectors
     for (int sq = 0; sq < config_.num_subquantizers; ++sq) {
         const size_t sq_index = static_cast<size_t>(sq);
-        if (sq_index >= static_cast<int>(codebooks_.size())) {
+        if (sq_index >= codebooks_.size()) {
             THEMIS_ERROR("ProductQuantizer::decode - Subquantizer index out of range");
             return {};
         }
@@ -340,12 +340,12 @@ float ProductQuantizer::computeAsymmetricDistance(
         return std::numeric_limits<float>::max();
     }
     
-    if (static_cast<int>(query.size()) != static_cast<size_t>(dimension_)) {
+    if (query.size() != static_cast<size_t>(dimension_)) {
         THEMIS_ERROR("ProductQuantizer::computeAsymmetricDistance - Query dimension mismatch");
         return std::numeric_limits<float>::max();
     }
     
-    if (static_cast<int>(codes.size()) != static_cast<size_t>(config_.num_subquantizers)) {
+    if (codes.size() != static_cast<size_t>(config_.num_subquantizers)) {
         THEMIS_ERROR("ProductQuantizer::computeAsymmetricDistance - Code size mismatch");
         return std::numeric_limits<float>::max();
     }
@@ -382,7 +382,7 @@ float ProductQuantizer::computeAsymmetricDistance(
         for (int i = 0; i < config_.num_subquantizers; ++i) {
             const size_t table_index =
                 static_cast<size_t>(i) * static_cast<size_t>(config_.num_centroids) + codes[static_cast<size_t>(i)];
-            if (table_index >= static_cast<int>(dis_table.size())) {
+            if (table_index >= dis_table.size()) {
                 THEMIS_ERROR("ProductQuantizer::computeAsymmetricDistance - Distance table index out of range");
                 return std::numeric_limits<float>::max();
             }
@@ -455,7 +455,7 @@ std::vector<std::vector<float>> ProductQuantizer::runKMeans(
     }
 
     for (const auto& subvector : subvector_data) {
-        if (static_cast<int>(subvector.size()) != static_cast<size_t>(subvector_dim_)) {
+        if (subvector.size() != static_cast<size_t>(subvector_dim_)) {
             THEMIS_WARN("ProductQuantizer::runKMeans - Subvector dimension mismatch");
             return {};
         }
@@ -552,7 +552,7 @@ std::vector<std::vector<float>> ProductQuantizer::runKMeans(
         // Sample next centroid with probability proportional to D^2
         std::discrete_distribution<uint64_t> weighted_dis(distances.begin(), distances.end());
         const size_t selected_index = static_cast<size_t>(weighted_dis(gen));
-        if (selected_index >= static_cast<int>(subvector_data.size())) {
+        if (selected_index >= subvector_data.size()) {
             THEMIS_WARN("ProductQuantizer::runKMeans - Weighted centroid index out of range");
             return {};
         }
@@ -610,7 +610,7 @@ std::vector<std::vector<float>> ProductQuantizer::runKMeans(
             } else {
                 // Handle empty cluster: reinitialize
                 const size_t reset_index = dis(gen);
-                if (reset_index >= static_cast<int>(subvector_data.size())) {
+                if (reset_index >= subvector_data.size()) {
                     THEMIS_WARN("ProductQuantizer::runKMeans - Reinitialization index out of range");
                     return {};
                 }
@@ -649,8 +649,8 @@ uint8_t ProductQuantizer::findNearestCentroid(
 }
 
 float ProductQuantizer::l2Distance(const std::vector<float>& a, const std::vector<float>& b) {
-    if (static_cast<int>(a.size()) != static_cast<int>(b.size())) {
-        THEMIS_DEBUG("ProductQuantizer::l2Distance - vector size mismatch ({} != {})",static_cast<int>(a.size()),static_cast<int>(b.size()));
+    if (a.size() != b.size()) {
+        THEMIS_DEBUG("ProductQuantizer::l2Distance - vector size mismatch ({} != {})", a.size(), b.size());
         return std::numeric_limits<float>::max();
     }
