+++ src/observability/metrics_collector.cpp
@@ -25,7 +25,7 @@ namespace observability {
 namespace {
 
 std::string sanitizeDiagnosticLabelValue(const std::string& value) {
-    if (static_cast<int>(value.size()) <= kMaxLabelValueBytes) {
+    if (value.size() <= static_cast<size_t>(kMaxLabelValueBytes)) {
         return value;
     }
     return value.substr(0, kMaxLabelValueBytes);
@@ -496,7 +496,7 @@ std::string MetricsCollector::formatExemplar(const Exemplar& exemplar) {
 
 bool MetricsCollector::areLabelsValid(const std::map<std::string, std::string>& labels,
                                       std::string* failure_reason) {
-    if (static_cast<int>(labels.size()) > kMaxMetricLabels) {
+    if (labels.size() > static_cast<size_t>(kMaxMetricLabels)) {
         if (failure_reason != nullptr) {
             *failure_reason = "label_count_exceeded";
         }
@@ -504,13 +504,13 @@ bool MetricsCollector::areLabelsValid(const std::map<std::string, std::string>& 
     }
 
     for (const auto& [key, value] : labels) {
-        if (static_cast<int>(key.size()) > kMaxLabelKeyBytes) {
+        if (key.size() > static_cast<size_t>(kMaxLabelKeyBytes)) {
             if (failure_reason != nullptr) {
                 *failure_reason = "label_key_too_long";
             }
             return false;
         }
-        if (static_cast<int>(value.size()) > kMaxLabelValueBytes) {
+        if (value.size() > static_cast<size_t>(kMaxLabelValueBytes)) {
             if (failure_reason != nullptr) {
                 *failure_reason = "label_value_too_long";
             }
@@ -527,8 +527,8 @@ void MetricsCollector::Histogram::observe(double value) {
     values.push_back(value);
     
     // Keep only recent samples
-    if (static_cast<int>(values.size()) > max_samples) {
-        values.erase(values.begin(), values.begin() + (static_cast<int>(values.size()) - max_samples));
+    if (values.size() > static_cast<size_t>(max_samples)) {
+        values.erase(values.begin(), values.begin() + static_cast<std::ptrdiff_t>(values.size() - static_cast<size_t>(max_samples)));
     }
 }
 
@@ -546,7 +546,7 @@ double MetricsCollector::Histogram::percentile(double p) const {
     std::vector<double> sorted = values;
     std::sort(sorted.begin(), sorted.end());
     
-    size_t index = static_cast<size_t>(p * (static_cast<int>(sorted.size()) - 1));
+    size_t index = static_cast<size_t>(p * static_cast<double>(sorted.size() - 1));
     return sorted[index];
 }
 
@@ -576,4 +576,3 @@ double LatencyTracker::elapsedMs() const {
 
 } // namespace observability
 } // namespace themis
-
